Move all mocks and types to @app/queries/*
