+++ src/app/Mappings/components/MappingBuilder/MappingBuilder.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Button, TextContent, Text, Grid, GridItem, Bullseye, Flex } from '@patternfly/react-core';
 import { PlusCircleIcon, TrashIcon } from '@patternfly/react-icons';
 import spacing from '@patternfly/react-styles/css/utilities/Spacing/spacing';
-import { MappingType, MappingSource, MappingTarget } from '../../types';
+import { MappingType, MappingSource, MappingTarget } from '@app/queries/types';
 import LineArrow from '@app/common/components/LineArrow';
 import MappingSourceSelect from './MappingSourceSelect';
 import MappingTargetSelect from './MappingTargetSelect';
