+++ src/app/Plans/components/Wizard/FilterVMsForm.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Form, FormGroup, TreeView, Title } from '@patternfly/react-core';
 
 // TODO replace with redux or equiv.
-import { VMsOptions } from './mocks/VMs.mock';
+import { VMsOptions } from '@app/queries/mocks/vms.mock';
 
 interface IFilterVMsProps {
   Inventory: [];
