--- src/app/Plans/mocks/plans.mock.ts
@@ -1,3 +1,0 @@
-import { IPlan } from '../types';
-
-export const MOCK_PLANS: IPlan[] = [];
