+++ src/app/queries/mocks/mappings.mock.ts
@@ -1,7 +1,7 @@
-import { IStorageMapping, INetworkMapping, MappingType } from '../types';
+import { IStorageMapping, INetworkMapping, MappingType } from '../types/mappings.types';
 import { ProviderType } from '@app/common/constants';
-import { MappingSource, MappingTarget } from '@app/Mappings/types';
-import { NetworkType } from '@app/Providers/types';
+import { MappingSource, MappingTarget } from '../types/mappings.types';
+import { NetworkType } from '../types/providers.types';
 
 const storageMapping1: IStorageMapping = {
   type: MappingType.Storage,
