+++ src/app/queries/mocks/plans.mock.ts
@@ -1,0 +1,3 @@
+import { IPlan } from '../types/plans.types';
+
+export const MOCK_PLANS: IPlan[] = [];
