+++ src/app/queries/types/common.types.ts
