+++ src/app/queries/types/hooks.types.ts
