+++ src/app/queries/types/providers.types.ts
